Workaround for #117
When using the baikal:nginx docker image in a subdirectory behind a
reverse proxy, the baikal base_uri setting doesn't work properly.
Rewriting the URI in the reverse proxy doesn't work nicely, so I fixed
it in the nginx config. Now the well-known URIs and the admin interface
seem to work, as well as the DAV endpoints.

Since this change hardcodes the subdirectory, it can't be upstreamed,
but may serve as an example to work around the issue.

Note you also need `base_uri: /dav/` in baikal.yaml.